+++ test/test_sock.sh
@@ -72,10 +72,10 @@ EOF
     expectpart "$($clixon_cli -1f $cfg show version)" 0 "${CLIXON_VERSION}"
  
     new "hello session-id 2"
-    expecteof "$clixon_util_socket -a $family -s $sock -D $DBG" 0 "<hello $DEFAULTNS/>" "<hello $DEFAULTNS><session-id>3</session-id></hello>"
+    expecteof "$clixon_util_socket -a $family -s $sock -D $DBG" 0 "<hello $DEFAULTONLY/>" "<hello $DEFAULTONLY><session-id>3</session-id></hello>"
 
     new "hello session-id 2"
-    expecteof "$clixon_util_socket -a $family -s $sock -D $DBG" 0 "<hello $DEFAULTNS/>" "<hello $DEFAULTNS><session-id>4</session-id></hello>"
+    expecteof "$clixon_util_socket -a $family -s $sock -D $DBG" 0 "<hello $DEFAULTONLY/>" "<hello $DEFAULTONLY><session-id>4</session-id></hello>"
 
     if [ $BE -ne 0 ]; then
 	new "Kill backend"